+++ ChangeLog.md
@@ -14,6 +14,7 @@
 New features:
 
 - [KT-11768](https://youtrack.jetbrains.com/issue/KT-11768) Implement "Introduce local variable" intention
+- [KT-11692](https://youtrack.jetbrains.com/issue/KT-11692) Support Spring model diagrams for Kotlin classes
 
 Issues fixed: